undrill remove workspace
